prepare user admin solutions

+++ src/main/java/com/codecool/stackoverflowtw/dao/UsersDaoJdbc.java
@@ -17,6 +17,7 @@ public class UsersDaoJdbc implements UsersDAO {
         this.database = database;
     }
 
+
     @Override
     public List<User> getAllUser() {
         String getAllUsers = "SELECT * FROM users";
@@ -36,8 +37,7 @@ public class UsersDaoJdbc implements UsersDAO {
     public User getUserById(int userId) {
         String userById = "SELECT * FROM users WHERE users.user_id = ?";
         User user = null;
-        try (Connection connection = database.getConnection();
-             PreparedStatement statement = connection.prepareStatement(userById)) {
+        try (Connection connection = database.getConnection(); PreparedStatement statement = connection.prepareStatement(userById)) {
             statement.setInt(1, userId);
             ResultSet resultSet = statement.executeQuery();
             if (resultSet.next()) {
@@ -57,26 +57,31 @@ public class UsersDaoJdbc implements UsersDAO {
 
     @Override
     public boolean deleteUser(int userId) {
-        String deleteUser = "DELETE FROM users WHERE id = ?";
-        try (Connection connection = database.getConnection(); PreparedStatement statement = connection.prepareStatement(deleteUser)) {
-            statement.setInt(1, userId);
-            int rowsDeleted = statement.executeUpdate();
-            System.out.println("User deleted. :) user_id : " + userId + ".");
-            return rowsDeleted > 0;
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
+        User userForDelete = getAllUser().stream().filter(user -> user.getUserId() == userId).findAny().orElseThrow();
+        if (userForDelete.isAdmin() && getAllUser().stream().filter(user -> user.isAdmin()).count() > 1 || !userForDelete.isAdmin()) {
+            String deleteUser = "DELETE FROM users WHERE id = ?";
+            try (Connection connection = database.getConnection(); PreparedStatement statement = connection.prepareStatement(deleteUser)) {
+                statement.setInt(1, userId);
+                int rowsDeleted = statement.executeUpdate();
+                System.out.println("User deleted. :) user_id : " + userId + ".");
+                return rowsDeleted > 0;
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
+        return false;
     }
 
     @Override
-    public void addUser(String userName, String userPassword) {
+    public void addUser(String userName, String userPassword, String email) {
         Date date = new Date();
-        User newUser = new User(userName, new Timestamp(date.getTime()), userPassword);
+        User newUser = new User(userName, userPassword, email, new Timestamp(date.getTime()));
+
         post(newUser);
     }
 
     public void post(User user) {
-        String template = "INSERT INTO users(id, user_name,registration_date, password, is_admin) values(DEFAULT,?,?,?,?) ";
+        String template = "INSERT INTO users(id, user_name,password, user_email, registration_date,  is_admin, is_rejected) values(DEFAULT,?,?,?,?,?,?) ";
         try (Connection connection = database.getConnection(); PreparedStatement statement = connection.prepareStatement(template)) {
             prepare(user, statement);
             statement.executeUpdate();
@@ -87,18 +92,17 @@ public class UsersDaoJdbc implements UsersDAO {
     }
 
     private User toEntity(ResultSet resultSet) throws SQLException {
-        return new User(resultSet.getInt("id"), resultSet.getString("user_name"),
-                resultSet.getTimestamp("registration_date"),
-                resultSet.getString("password"),
-                resultSet.getBoolean("is_admin"));
+        return new User(resultSet.getInt("id"), resultSet.getString("user_name"), resultSet.getString("password"), resultSet.getString("user_email"), resultSet.getTimestamp("registration_date"), resultSet.getBoolean("is_admin"));
     }
 
     private void prepare(User user, PreparedStatement statement) throws SQLException {
         /* statement.setInt(1, user.getUserId());*/
         statement.setString(1, user.getUserName());
-        statement.setTimestamp(2, user.getRegistrationDateTime());
-        statement.setString(3, user.getPassword());
-        statement.setBoolean(4, user.isChecked());
+        statement.setString(2, user.getPassword());
+        statement.setString(3, user.getUserEmail());
+        statement.setTimestamp(4, user.getRegistrationDateTime());
+        statement.setBoolean(5, user.isAdmin());
+        statement.setBoolean(6, user.isRejected());
 
     }
 }
